requirements.txt
google-api-core==1.28.0
google-api-python-client==2.6.0
google-auth==1.30.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
python-dotenv==0.17.1